# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: "Medical Triage Admin Portal — Express REST API"
  description: |
    REST API for semantic triage case search and vector index management.

    This API powers the Admin Portal's vector search features and the
    file ingestion pipeline (server/services/fileWatcher.ts). It runs
    alongside the React SPA on the Express backend.

    ## Authentication

    All `/api/vector/*` routes are protected by Bearer token auth.
    Set `ADMIN_API_KEY` in `.env` to enable authentication. Without it,
    all requests are allowed in local dev mode.

    ## Import into Postman / Insomnia

    - **Postman**: File → Import → select this file
    - **Insomnia**: Application → Preferences → Import Data
    - **Swagger UI**: Point to `/api-docs` (not yet deployed)

    The base URL defaults to `http://localhost:5001`. Change the
    `servers[0].url` value for staging/production.

    ## Related Documentation

    - Integration audit checklist: `docs/INTEGRATION_CHECKLIST.md`
    - LLM inference contract: `docs/LLM_TRIAGE_API_CONTRACT.md`
    - Server source: `server/index.ts`
    - Routes: `server/routes/vectorRoutes.ts`

  version: "1.0.0"
  contact:
    name: "Admin Portal Team"
servers:
  - url: "http://localhost:5001"
    description: "Local development server"
  - url: "https://api.apexhealth.internal"
    description: "Staging/Production (update URL)"

# ── Security Schemes ──
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API_KEY
      description: |
        Shared secret set via `ADMIN_API_KEY` in `.env`.
        If `ADMIN_API_KEY` is not configured, auth is skipped (local dev mode).
    Unauthenticated:
      type: apiKey
      in: header
      name: X-No-Auth
      description: "No authentication required (health/ready probes only)"

  # ── Reusable Schemas ──
  schemas:
...
    ErrorResponse:
      type: object
      description: "Standard error payload returned by all endpoints."
      properties:
        error:
          type: string
          description: "Machine-readable error code."
          example: "INVALID_ARGUMENT"
        message:
          type: string
          description: "Human-readable error description."
          example: "'query' must be a non-empty string."
      required:
        - error
        - message
    ServerError:
      type: object
      properties:
        error:
          type: string
          example: "INTERNAL_SERVER_ERROR"
        message:
          type: string
          example: "An internal error occurred."
...
    IndexSingleRequest:
      type: object
      properties:
        id:
          type: string
          example: "CASE-1001"
        text:
          type: string
          example: "Patient: Michael Chen\nTriage Category: Emergency"
        metadata:
          type: object
          properties:
            patientName:
              type: string
            triageCategory:
              type: string
            status:
              type: string
            confidence:
              type: number
            timestamp:
              type: string
              format: date-time
            sourceType:
              type: string
              enum: [patient, report]
      required:
        - id
        - text
    IndexSingleResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true

    # ── Delete Vectors ──
    DeleteVectorsRequest:
      type: object
      properties:
        ids:
          type: array
          description: "Vector IDs to delete from the Pinecone index."
          items:
            type: string
          example:
            - "CASE-1001"
            - "report_CASE-2001"
      required:
        - ids
    DeleteVectorsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        deletedCount:
          type: integer
          example: 2

    # ── Vector Stats ──
...
  responses:
    Unauthorized:
      description: "Missing or malformed Authorization header."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "UNAUTHORIZED"
            message: "Missing or malformed Authorization header. Expected: Bearer <API_KEY>"
    Forbidden:
      description: "Invalid API key."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "FORBIDDEN"
            message: "Invalid API key."
    InternalError:
      description: "Unexpected server error."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServerError"
          example:
            error: "INTERNAL_SERVER_ERROR"
            message: "An internal error occurred."

# ── API Paths ──
paths:
...
  /api/vector/index:
    post:
      tags:
        - "Index Management"
      summary: "Index a single triage case"
      description: |
        Indexes a single triage case into Pinecone.

        Generates an embedding from the provided text and upserts
        it to the Pinecone index with associated metadata.
      operationId: "indexVector"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IndexSingleRequest"
            example:
              id: "CASE-1001"
              text: "Patient: Michael Chen\nTriage Category: Emergency"
              metadata:
                patientName: "Michael Chen"
                triageCategory: "Emergency"
                status: "Escalated"
                confidence: 0.98
                timestamp: "2026-06-08T10:00:00.000Z"
                sourceType: "patient"
      responses:
        "200":
          description: "Vector indexed successfully."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndexSingleResponse"
        "400":
          description: "Invalid request — id or text is missing."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/vector:
    delete:
      tags:
        - "Index Management"
      summary: "Delete vectors by ID"
      description: "Deletes one or more vectors from the Pinecone index by their document IDs."
      operationId: "deleteVectors"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteVectorsRequest"
            example:
              ids:
                - "CASE-1001"
                - "report_CASE-2001"
      responses:
        "200":
          description: "Vectors deleted successfully."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteVectorsResponse"
              example:
                success: true
                deletedCount: 2
        "400":
          description: "Invalid request — ids is empty or not an array."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...